Extracted Versioned to own module, generalised its derivations.
