+++ libs/wire-api/wire-api.cabal
@@ -95,6 +95,7 @@ library
     Wire.API.Routes.QualifiedCapture
     Wire.API.Routes.Version
     Wire.API.Routes.Version.Wai
+    Wire.API.Routes.Versioned
     Wire.API.Routes.WebSocket
     Wire.API.ServantProto
     Wire.API.Swagger
